requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
altair==6.0.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anthropic==0.84.0
    # via langchain-anthropic
anyio==4.12.1
    # via
    #   anthropic
    #   httpx
    #   mcp
    #   openai
    #   sse-starlette
    #   starlette
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.6
    # via streamlit
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via
    #   pdfminer-six
    #   requests
click==8.3.1
    # via
    #   streamlit
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
cryptography==46.0.5
    # via
    #   pdfminer-six
    #   pyjwt
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via
    #   anthropic
    #   openai
docstring-parser==0.17.0
    # via anthropic
et-xmlfile==2.0.0
    # via openpyxl
faiss-cpu==1.13.2
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
greenlet==3.3.2
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   langgraph-sdk
    #   langsmith
    #   mcp
    #   openai
httpx-sse==0.4.3
    # via
    #   langchain-community
    #   mcp
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.13.0
    # via
    #   anthropic
    #   openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.26.0
    # via
    #   altair
    #   mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
langchain==1.2.10
    # via -r requirements.in
langchain-anthropic==1.3.4
    # via -r requirements.in
langchain-classic==1.0.1
    # via langchain-community
langchain-community==0.4.1
    # via -r requirements.in
langchain-core==1.2.16
    # via
    #   langchain
    #   langchain-anthropic
    #   langchain-classic
    #   langchain-community
    #   langchain-mcp-adapters
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-mcp-adapters==0.2.1
    # via -r requirements.in
langchain-openai==1.1.10
    # via -r requirements.in
langchain-text-splitters==1.1.1
    # via langchain-classic
langgraph==1.0.10
    # via
    #   -r requirements.in
    #   langchain
langgraph-checkpoint==4.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.8
    # via langgraph
langgraph-sdk==0.3.9
    # via langgraph
langsmith==0.7.9
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
lxml==6.0.2
    # via python-docx
markupsafe==3.0.3
    # via jinja2
marshmallow==3.26.2
    # via dataclasses-json
mcp==1.26.0
    # via
    #   -r requirements.in
    #   langchain-mcp-adapters
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
narwhals==2.17.0
    # via altair
numpy==2.4.2
    # via
    #   faiss-cpu
    #   langchain-community
    #   pandas
    #   pydeck
    #   streamlit
openai==2.24.0
    # via langchain-openai
openpyxl==3.1.5
    # via -r requirements.in
orjson==3.11.7
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.2
    # via langgraph-checkpoint
packaging==26.0
    # via
    #   altair
    #   faiss-cpu
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   streamlit
pandas==2.3.3
    # via streamlit
pdfminer-six==20251230
    # via pdfplumber
pdfplumber==0.11.9
    # via -r requirements.in
pillow==12.1.1
    # via
    #   pdfplumber
    #   streamlit
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.33.5
    # via streamlit
pyarrow==23.0.1
    # via streamlit
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   anthropic
    #   langchain
    #   langchain-anthropic
    #   langchain-classic
    #   langchain-core
    #   langgraph
    #   langsmith
    #   mcp
    #   openai
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via
    #   langchain-community
    #   mcp
pydeck==0.9.1
    # via streamlit
pyjwt[crypto]==2.11.0
    # via mcp
pypdfium2==5.5.0
    # via pdfplumber
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.2.0
    # via -r requirements.in
python-dotenv==1.2.2
    # via pydantic-settings
python-multipart==0.0.22
    # via mcp
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2026.2.28
    # via tiktoken
requests==2.32.5
    # via
    #   langchain-classic
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   streamlit
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anthropic
    #   openai
sqlalchemy==2.0.48
    # via
    #   langchain-classic
    #   langchain-community
sse-starlette==3.3.2
    # via mcp
starlette==0.52.1
    # via
    #   mcp
    #   sse-starlette
streamlit==1.54.0
    # via -r requirements.in
tenacity==9.1.4
    # via
    #   langchain-community
    #   langchain-core
    #   streamlit
tiktoken==0.12.0
    # via langchain-openai
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   altair
    #   anthropic
    #   anyio
    #   langchain-core
    #   langchain-mcp-adapters
    #   mcp
    #   openai
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   referencing
    #   sqlalchemy
    #   starlette
    #   streamlit
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
uuid-utils==0.14.1
    # via
    #   langchain-core
    #   langsmith
uvicorn==0.41.0
    # via mcp
watchdog==6.0.0
    # via streamlit
xxhash==3.6.0
    # via
    #   langgraph
    #   langsmith
yarl==1.23.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
